fixed wagmi issue
