+++ package.json
@@ -19,9 +19,6 @@
     "@types/node": "20.4.1",
     "@types/react": "^18.2.48",
     "@types/react-dom": "18.2.6",
-    "@walletconnect/ethereum-provider": "^2.10.0",
-    "@web3modal/ethereum": "^2.7.1",
-    "@web3modal/react": "^2.7.1",
     "@web3modal/wagmi": "^4.2.0-alpha.0",
     "add": "^2.0.6",
     "dotenv": "^16.3.1",
